+++ lbrynet/wallet/transaction.py
@@ -76,9 +76,7 @@ class Transaction(BaseTransaction):
     def pay(cls, amount: int, address: bytes, funding_accounts: List[Account], change_account: Account):
         ledger = cls.ensure_all_have_same_ledger(funding_accounts, change_account)
         hash_of_address = ledger.address_to_hash160(address)
-
         output = Output.pay_pubkey_hash(amount, hash_of_address)
-
         return cls.create([], [output], funding_accounts, change_account)
 
     @classmethod